[fix] The sound of buttons have been enabled.
git-svn-id: svn://www.hshgames.com/project_s@582 e3c12c2a-3f9b-4cee-bc1f-3a9786f174a7

+++ Classes/MyUI/Popup/Option/OptionMusicPopup.hpp
@@ -11,11 +11,11 @@ protected:
     
 private:
     COptionMusicPopup()
-    : m_BGMVolume(0)
-    , m_EffectVolume(0){};
+    : m_BGMVolume(0.f)
+    , m_EffectVolume(0.f){};
     virtual ~COptionMusicPopup(){};
     
 private:
-    int m_BGMVolume;
-    int m_EffectVolume;
+    float m_BGMVolume;
+    float m_EffectVolume;
 };
